Issue #1637 Don't run daemon on windows as service
While running the daemon process as a service, it always
runs as the administrator if the logged in user for the
service is also a local adminstrator, which we don't want
since our preflight checks prevents from executing as admin

As a workaround we are going to run the daemon process in the
background with a hidden windows, the next commits handles the
setup of the daemon to run as such and auto-start after reboot
